WIP #86 - Fix rather stupid problem with SpecFlow
This was that the default plugin type is a runtime/generator plugin, and
so it was searching for my plugin when I tried to save an edited file.

This in turn triggered other bugs - but it was all down to the fact that
it was looking for a generator plugin.