Resolve #23 by officially removing support for Edge 13.x
This browser does not work correctly even without the modifier key
pressed.  It appears that in Edge 14, Microsoft fixed a real bug
relating to HTML <select multiple> but they fixed it requiring a
modifier key.

This commit essentially reverts the work introduced in:

  191f7b2f2c56b4640573f2ab0a9e4d673aad5f00

Except that rather than leave that browser commented-out in the
build matrix, this commit outright removes it.