Resolve #33 - Safari 11 doesn't seem to be compatible with <select> elements
Oh dear.  I had thought that Safari 11 would require command+click in order
to toggle state of <select multiple> elements.  The same problems I am
seeing for multi-select elements, I am also seeing for single-select
elements.  That suggests that the problem actually has nothing to do with
select elements with the multiple attribute.

Right now, as far as I can tell, Safari v11 (specifically, it's webdriver)
is plain incompatible with anything that changes the state of a select
element.

I've set up a browser flag to indicate this.  If I'm lucky I'll be able to
find a better solution or workaround someday.