WIP #208 - Copy log values formatter from .NET
This (MIT licensed) content seems to do a lot of the
work I'd like to be doing in Screenplay for reporting.
Specifically, that's formatting templates which are
written like:
    "{Actor} washes {Count} dishes"
and turning those into formatted strings like:
    "Joe washes 5 dishes"
As you can see that's a lot like the String.Format
syntax but with named placeholders rather than
numbered ones.  This class seems to do that conversion
job in MS extensions logging, and so I've copied it so
that I can try it out and customize it myself to verify
whether it's what I want.

For the record, the source of this file is
    https://github.com/dotnet/runtime/blob/49399d99e11d356c412c9fd2e25af3435abe2e13/src/libraries/Microsoft.Extensions.Logging.Abstractions/src/LogValuesFormatter.cs