Go back to nunit 2.x endpoint
It seems that the issue using the nunit
3 endpoint persists. Look at the 
history of this file for a now-deleted 
comment that links to more info about 
that. It seems that after 6 months,
it's still not resolved. 

The symptom is uploading an nunit 3
test result file which includes test
durations expressed in scientific
notation causes HTTP 500 errors from
the Appveyor endpoint.